Изменены типы данных для нахождения значения функции в CalculateFunctionConverter

+++ FunctionsWpf/Infrastructure/Converters/CalculateFunctionConverter.cs
@@ -9,16 +9,16 @@ namespace FunctionsWpf.Infrastructure.Converters
     {
         public object Convert(object[] values, Type targetType, object parameter, CultureInfo culture)
         {
-            int a = int.Parse(values[0].ToString());
-            int b = int.Parse(values[1].ToString());
+            double a = double.Parse(values[0].ToString());
+            double b = double.Parse(values[1].ToString());
             int c = int.Parse(values[2].ToString());
-            int x = int.Parse(values[3].ToString());
-            int y = int.Parse(values[4].ToString());
+            double x = double.Parse(values[3].ToString());
+            double y = double.Parse(values[4].ToString());
             int functionIndex = int.Parse(values[5].ToString());
             Function.FunctionType functionType = Function.GetFunctionTypeFromIndex(functionIndex);
 
             Function function = new Function(x, y);
-            long functionValue = function.Calculate(a, b, c, functionType);
+            double functionValue = function.Calculate(a, b, c, functionType);
             return functionValue.ToString();
         }
